Mobile 1746

+++ www/addons/mod/assign/submission/file/handlers.js
@@ -14,6 +14,8 @@
 
 angular.module('mm.addons.mod_assign')
 
+.constant('mmaModAssignSubmissionFileName', 'submission_file')
+
 /**
  * Handler for file submission plugin.
  *
@@ -22,7 +24,7 @@ angular.module('mm.addons.mod_assign')
  * @name $mmaModAssignSubmissionFileHandler
  */
 .factory('$mmaModAssignSubmissionFileHandler', function($mmaModAssignSubmissionFileSession, $mmaModAssign, $mmSite, $q,
-            $mmaModAssignHelper, $mmWS, $mmFS, $mmFilepool, $mmUtil, mmaModAssignComponent) {
+            $mmaModAssignHelper, $mmWS, $mmFS, $mmFilepool, $mmUtil, $mmaModAssignOffline, mmaModAssignSubmissionFileName) {
 
     var self = {};
 
@@ -43,7 +45,7 @@ angular.module('mm.addons.mod_assign')
 
         // Now delete the local files from the tmp folder.
         files.forEach(function(file) {
-            if (file.remove) {
+            if (!file.offline && file.remove) {
                 file.remove();
             }
         });
@@ -64,6 +66,23 @@ angular.module('mm.addons.mod_assign')
 
         return $mmaModAssignHelper.uploadFiles(assign.id, files).then(function(itemId) {
             pluginData.files_filemanager = itemId;
+        });
+    };
+
+    /**
+     * Delete offline data stored for a certain submission.
+     *
+     * @param  {Object} assign      Assignment.
+     * @param  {Object} submission  Submission affected.
+     * @param  {Object} plugin      Plugin to delete the data for.
+     * @param  {Object} offlineData Offline data stored for the submission.
+     * @param  {String} [siteId]    Site ID. If not defined, current site.
+     * @return {Promise}            Promise resolved when deleted.
+     */
+    self.deleteOfflineData = function(assign, submission, plugin, offlineData, siteId) {
+        return $mmaModAssignHelper.deleteStoredSubmissionFiles(assign.id,
+                mmaModAssignSubmissionFileName, submission.userid, siteId).catch(function() {
+            // Ignore errors, maybe the folder doesn't exist.
         });
     };
 
@@ -113,7 +132,7 @@ angular.module('mm.addons.mod_assign')
                 promises = [];
 
             angular.forEach(files, function(file) {
-                if (file.filename) {
+                if (file.filename && !file.name) {
                     // It's a remote file. First check if we have the file downloaded since it's more reliable.
                     promises.push($mmFilepool.getFilePathByUrl(siteId, file.fileurl).then(function(path) {
                         return $mmFS.getFile(path).then(function(fileEntry) {
@@ -203,38 +222,54 @@ angular.module('mm.addons.mod_assign')
      * @return {Promise}           Promise resolved with true if data has changed, resolved with false otherwise.
      */
     self.hasDataChanged = function(assign, submission, plugin, inputData) {
-        var currentFiles = $mmaModAssignSubmissionFileSession.getFiles(assign.id),
-            initialFiles = $mmaModAssign.getSubmissionPluginAttachments(plugin);
+        // Check if there's any offline data.
+        return $mmaModAssignOffline.getSubmission(assign.id, submission.userid).catch(function() {
+                // No offline data found.
+        }).then(function(offlineData) {
+            if (offlineData && offlineData.plugindata && offlineData.plugindata.files_filemanager) {
+                // Has offline data, return the number of files.
+                return offlineData.plugindata.files_filemanager.offline + offlineData.plugindata.files_filemanager.online.length;
+            }
+            // No offline data, return the number of online files.
+            var pluginFiles = $mmaModAssign.getSubmissionPluginAttachments(plugin);
+            return pluginFiles && pluginFiles.length;
+        }).then(function(numFiles) {
+            var currentFiles = $mmaModAssignSubmissionFileSession.getFiles(assign.id);
 
-        if (currentFiles.length != initialFiles.length) {
-            return true;
-        }
-
-        // Search if there is any local file added.
-        for (var i = 0; i < currentFiles.length; i++) {
-            var file = currentFiles[i];
-            if (!file.filename && typeof file.name != 'undefined') {
-                // There's a local file added, list has changed.
+            if (currentFiles.length != numFiles) {
+                // Number of files has changed.
                 return true;
             }
-        }
 
-        // No local files and list length is the same, this means the list hasn't changed.
-        return false;
+            // Search if there is any local file added.
+            for (var i = 0; i < currentFiles.length; i++) {
+                var file = currentFiles[i];
+                if (!file.filename && typeof file.name != 'undefined' && !file.offline) {
+                    // There's a local file added, list has changed.
+                    return true;
+                }
+            }
+
+            // No local files and list length is the same, this means the list hasn't changed.
+            return false;
+        });
     };
 
     /**
      * Should prepare and add to pluginData the data to send to server based in the input data.
      *
      * @param  {Object} assign     Assignment.
-     * @param  {Object} submission Submission to check data.
+     * @param  {Object} submission Submission affected.
      * @param  {Object} plugin     Plugin to get the data for.
      * @param  {Object} inputData  Data entered in the submission form.
      * @param  {Object} pluginData Object where to add the plugin data.
-     * @return {Void}
+     * @param  {Boolean} offline   True to prepare the data for an offline submission, false otherwise.
+     * @param  {Number} [userId]   User ID. If not defined, site's current user.
+     * @param  {String} [siteId]   Site ID. If not defined, current site.
+     * @return {Promise|Void}
      */
-    self.prepareSubmissionData = function(assign, submission, plugin, inputData, pluginData) {
-        var siteId = $mmSite.getId();
+    self.prepareSubmissionData = function(assign, submission, plugin, inputData, pluginData, offline, userId, siteId) {
+        siteId = siteId || $mmSite.getId();
 
         if (self.hasDataChanged(assign, submission, plugin, inputData)) {
             // Data has changed, we need to upload new files and re-upload all the existing files.
@@ -245,8 +280,47 @@ angular.module('mm.addons.mod_assign')
                 return $q.reject(error);
             }
 
-            return $mmaModAssignHelper.uploadFiles(assign.id, currentFiles, siteId).then(function(itemId) {
-                pluginData.files_filemanager = itemId;
+            return $mmaModAssignHelper.uploadOrStoreFiles(assign.id, mmaModAssignSubmissionFileName,
+                        currentFiles, offline, userId, siteId).then(function(result) {
+                pluginData.files_filemanager = result;
+            });
+        }
+    };
+
+    /**
+     * Should prepare and add to pluginData the data to send to server to synchronize an offline submission.
+     *
+     * @param  {Object} assign      Assignment.
+     * @param  {Object} submission  Submission affected.
+     * @param  {Object} plugin      Plugin to get the data for.
+     * @param  {Object} offlineData Offline data stored for the submission.
+     * @param  {Object} pluginData  Object where to add the plugin data.
+     * @param  {String} [siteId]    Site ID. If not defined, current site.
+     * @return {Promise|Void}
+     */
+    self.prepareSyncData = function(assign, submission, plugin, offlineData, pluginData, siteId) {
+        var filesData = offlineData && offlineData.plugindata && offlineData.plugindata.files_filemanager;
+        if (filesData) {
+            // Has some data to sync.
+            var files = filesData.online || [],
+                promise;
+
+            if (filesData.offline) {
+                // Has offline files.
+                promise = $mmaModAssignHelper.getStoredSubmissionFiles(assign.id,
+                            mmaModAssignSubmissionFileName, submission.userid, siteId).then(function(result) {
+                    files = files.concat(result);
+                }).catch(function() {
+                    // Folder not found, no files to add.
+                });
+            } else {
+                promise = $q.when();
+            }
+
+            return promise.then(function() {
+                return $mmaModAssignHelper.uploadFiles(assign.id, files, siteId).then(function(itemId) {
+                    pluginData.files_filemanager = itemId;
+                });
             });
         }
     };
@@ -256,7 +330,7 @@ angular.module('mm.addons.mod_assign')
 
 .run(function($mmAddonManager) {
     // Use addon manager to inject $mmaModAssignSubmissionDelegate. This is to provide an example for remote addons,
-    // since they cannot assume that the quiz addon will be packaged in custom apps.
+    // since they cannot assume that the assign addon will be packaged in custom apps.
     var $mmaModAssignSubmissionDelegate = $mmAddonManager.get('$mmaModAssignSubmissionDelegate');
     if ($mmaModAssignSubmissionDelegate) {
         $mmaModAssignSubmissionDelegate.registerHandler('mmaModAssignSubmissionFile', 'file',
